fix: move getDayjsvalue into util
