+++ packages/statistic/src/util.ts
@@ -1,4 +1,6 @@
-// https://github.com/ant-design/ant-design/blob/master/components/statistic/utils.tsx
+import { isDayjs } from "@illa-design/system"
+import * as dayjs from "dayjs"
+
 const units: [string, number][] = [
   ["Y", 1000 * 60 * 60 * 24 * 365], // years
   ["M", 1000 * 60 * 60 * 24 * 30], // months
@@ -9,6 +11,7 @@ const units: [string, number][] = [
   ["S", 1], // million seconds
 ]
 
+// thx ant
 export function getDateString(millisecond: number, format: string) {
   let leftMillisecond: number = millisecond
   return units.reduce((current, [name, unit]) => {
@@ -22,4 +25,28 @@ export function getDateString(millisecond: number, format: string) {
     }
     return current
   }, format)
+}
+
+// thx arco
+export function getDayjsValue(time: any, format: string) {
+  if (!time) {
+    return void 0
+  }
+  const formatValue = (value: any) => {
+    if (isDayjs(value)) {
+      return dayjs(value.valueOf())
+    }
+    if (typeof value === "string") {
+      const dv = dayjs(value, format)
+
+      return dv.isValid() ? dv : dayjs(value, "YYYY-MM-DD")
+    }
+    return dayjs(value, format)
+  }
+
+  if (Array.isArray(time)) {
+    return time.map((t) => (t ? formatValue(t) : void 0))
+  }
+
+  return formatValue(time)
 }
