+++ packages/system/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./omit"
 export * from "./useMergeValue"
 export * from "./is"
-export * from "./getDayjsValue"
